am263x/am263px/am261x/am243x/am64x: bootloader: update struct fields
- update secure boot struct fields so as to reduce
  memory footprint of the data structure

Fixes: TIFSMCU-2254

+++ source/drivers/bootloader/soc/am261x/bootloader_soc.h
@@ -302,7 +302,7 @@ int32_t Bootloader_authStart(uintptr_t startAddr, uint32_t size);
 /**
  * \brief API to update the Secure Boot Stream with a new segment
  */
-int32_t Bootloader_authUpdate(uintptr_t startAddr, uint32_t size, uint32_t enc);
+int32_t Bootloader_authUpdate(uintptr_t startAddr, uint32_t size, uint8_t enc);
 
 /**
  * \brief API to finish streaming boot authentication
